Many new updates and unittests updates.
- Make modules CMakeInclude platform independent.
- Disabled pushbullet and pushover due to lack of
information to port to rewrite.
- Temporarily disabled DHT11 and DHT22 until ported.
- Temporarily disabled ds18b20 and ds18s20 until ported.
- Temporarily disabled the gpio_switch protocol until ported.
- Temporarily disabled ARPing and Program protocol until ported.
- Temporarily disabled Ping protocol until ported.
- Event parsing is now weakly typed by casting variables
between variable types. The IS operator is therefor dropped
and replaced by the type idependent == operator.
- Added unit tests for i2c protocols and ported them to rewrite.
- Some fixes in the http, mail, openweathermap, wunderground, and
xbmc unittest.
- Adapted event unit tests for new weakly types code.

+++ libs/pilight/core/dso.h
@@ -12,7 +12,7 @@
 typedef struct dso_t {
 	void *handle;
 	char *name;
-	ssize_t size;
+	unsigned long size;
 	struct dso_t *next;
 } dso_t;
